Rules: allow overriding with more specific title
Tags in rules are no longer additive, but each rule overrides the next.
Windows with a more specific (longer) title can override other windows
of the same program. So for example, windows with KeePassXC in the title
can be tiled and set to tag 7, but windows with "AutoType - KeePassXC"
(a more specific/longer match) can be configured as floating and
displayed on the current tag.